Moved dashboards pages -> components
